refactor(onchain-indexing): :recycle: extract conditional order processing out
